requirements.txt
aws-cdk-lib
constructs==10.4.2
python-dotenv